CmdLet: Invoke-ExpressionAt // Spread-Expression
Executes an expression at a given location, failing if moving to that
location is not possible. Inspects $? and $LASTEXITCODE to determine if
the expression succeeded and generates a non-terminating error if not.
Target path(s) may be provided as cli argument or via the pipeline.

Spread-Expression reads target paths from a line-based descriptor file
(default './.spread'). Each expression received as pipeline input is
executed in each of the targetted paths.