Install-PyProfile now supports relocation
Introduce -SourceBase and -TargetBase parameters in Install-PyProfile.ps1,
setting the base path of the expanded distribution and target installation
path respectively.
To prevent bad installations, installer will now skipping module
installation if the default module path is not existing. A warning is
emitted in that case.

Add unit test coverage for installer script.