Modular profile loading
Add the loader implementation as default profile file to installer module.

Already existing profile files are backed up to the new profile fragment
load path, using the default file name. The installation routine detects
if a profile is already a loader implementation and skips backing it up in
that case.

Installation of the profile loader requires confirmation if an existing
profile file is detected.